+++ changelog.md
@@ -10,6 +10,11 @@ TODO:
 
 
 
+*	2020.3.28
+
+	+	app/components/notations-matcher.vue: created.
+
+
 *	2020.3.27
 
 	+	inc/logRecorder.ts created.